+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 
+- Changed HATEOAS attribute and link authorization: both permission and policy must now be valid (previously either
+  condition was sufficient).
+
 ### Deprecated
 
 ### Removed